Add JSON scalar and mmessage-content typings
This sets up the GraphQL typings and resolvers for a lot of the message-content related API surface. A message is someone (a user) sending something (content). Content has a type and a data JSON field supported via Postgres. The idea is that the frontend will receive the JSON and the type and then will have a set of UI rules on how to draw messages of certain types. The types are things such as "text", or "poll", so that rendering this kind of dynamic material will be quite easy and organized. Also easily pluggable by simply extending the number of message content types you support internally.

To support the graphql api returning JSON we can extend the schema by adding a scalar JSON type that will have a custom resolver function. The resolver functionality is provided by apollo in graphql-type-json which you can expose as JSON in your resolver map